Separate MergingIterator and SsTableIterator from compaction module
