+++ src/core/mod.rs
@@ -1,6 +1,7 @@
 use std::cmp;
 
 pub mod storage;
+pub mod merge;
 pub use storage::{EngineError, StorageEngine};
 
 #[derive(thiserror::Error, Debug)]
